"LA.QSSI.11.0.r1-23500-qssi.0"

* tag 'LA.QSSI.11.0.r1-23500-qssi.0' of https://git.codelinaro.org/clo/la/platform/frameworks/base: (166 commits)
  [DO NOT MERGE] Enforce persisted snoozed notifications limits
  [RESTRICT AUTOMERGE] Check permission of Autofill icon URIs
  Restrict activity launch when caller is running in the background
  Unbind TileService onNullBinding
  DO NOT MERGE: "Hide" /Android/data|obb|sanbox/ on shared storage
  DO NOT MERGE Disallow Wallpaper service to launch activity from background.
  Fix vulnerability that allowed attackers to start arbitary activities
  RESTRICT AUTOMERGE Log to detect usage of whitelistToken when sending non-PI target
  [CDM] Validate component name length before requesting notification access.
  Truncate user data to a limit of 500 characters
  DO NOT MERGE Ensure finish lockscreen when usersetup incomplete
  Require permission to unlock keyguard
  RESTRICT AUTOMERGE: Check URI permissions for resumable media artwork
  Updated: always show the keyguard on device lockdown
  Adding in verification of calling UID in onShellCommand
  Revert "On device lockdown, always show the keyguard"
  Validate userId when publishing shortcuts
  Use readUniqueFileDescriptor in incidentd service
  Restrict number of shortcuts can be added through addDynamicShortcuts
  Validate URI-based shortcut icon at creation time.
  ...

+++ packages/SettingsProvider/src/com/android/providers/settings/SettingsProvider.java
@@ -1781,6 +1781,9 @@ public class SettingsProvider extends ContentProvider {
             cacheName = Settings.System.ALARM_ALERT_CACHE;
         }
         if (cacheName != null) {
+            if (!isValidAudioUri(name, value)) {
+                return false;
+            }
             final File cacheFile = new File(
                     getRingtoneCacheDir(owningUserId), cacheName);
             cacheFile.delete();
@@ -1811,6 +1814,34 @@ public class SettingsProvider extends ContentProvider {
 
             return false;
         }
+    }
+
+    private boolean isValidAudioUri(String name, String uri) {
+        if (uri != null) {
+            Uri audioUri = Uri.parse(uri);
+            if (Settings.AUTHORITY.equals(
+                    ContentProvider.getAuthorityWithoutUserId(audioUri.getAuthority()))) {
+                // Don't accept setting the default uri to self-referential URIs like
+                // Settings.System.DEFAULT_RINGTONE_URI, which is an alias to the value of this
+                // setting.
+                return false;
+            }
+            final String mimeType = getContext().getContentResolver().getType(audioUri);
+            if (mimeType == null) {
+                Slog.e(LOG_TAG,
+                        "mutateSystemSetting for setting: " + name + " URI: " + audioUri
+                        + " ignored: failure to find mimeType (no access from this context?)");
+                return false;
+            }
+            if (!(mimeType.startsWith("audio/") || mimeType.equals("application/ogg")
+                    || mimeType.equals("application/x-flac"))) {
+                Slog.e(LOG_TAG,
+                        "mutateSystemSetting for setting: " + name + " URI: " + audioUri
+                        + " ignored: associated mimeType: " + mimeType + " is not an audio type");
+                return false;
+            }
+        }
+        return true;
     }
 
     private boolean hasWriteSecureSettingsPermission() {
@@ -2955,6 +2986,15 @@ public class SettingsProvider extends ContentProvider {
             return settingsState.getSettingLocked(name);
         }
 
+        private boolean shouldExcludeSettingFromReset(Setting setting, String prefix) {
+            // If a prefix was specified, exclude settings whose names don't start with it.
+            if (prefix != null && !setting.getName().startsWith(prefix)) {
+                return true;
+            }
+            // Never reset SECURE_FRP_MODE, as it could be abused to bypass FRP via RescueParty.
+            return Secure.SECURE_FRP_MODE.equals(setting.getName());
+        }
+
         public void resetSettingsLocked(int type, int userId, String packageName, int mode,
                 String tag) {
             resetSettingsLocked(type, userId, packageName, mode, tag, /*prefix=*/
@@ -2977,7 +3017,7 @@ public class SettingsProvider extends ContentProvider {
                         Setting setting = settingsState.getSettingLocked(name);
                         if (packageName.equals(setting.getPackageName())) {
                             if ((tag != null && !tag.equals(setting.getTag()))
-                                    || (prefix != null && !setting.getName().startsWith(prefix))) {
+                                    || shouldExcludeSettingFromReset(setting, prefix)) {
                                 continue;
                             }
                             if (settingsState.resetSettingLocked(name)) {
@@ -2997,7 +3037,7 @@ public class SettingsProvider extends ContentProvider {
                         Setting setting = settingsState.getSettingLocked(name);
                         if (!SettingsState.isSystemPackage(getContext(),
                                 setting.getPackageName(), INVALID_UID, userId)) {
-                            if (prefix != null && !setting.getName().startsWith(prefix)) {
+                            if (shouldExcludeSettingFromReset(setting, prefix)) {
                                 continue;
                             }
                             if (settingsState.resetSettingLocked(name)) {
@@ -3017,7 +3057,7 @@ public class SettingsProvider extends ContentProvider {
                         Setting setting = settingsState.getSettingLocked(name);
                         if (!SettingsState.isSystemPackage(getContext(),
                                 setting.getPackageName(), INVALID_UID, userId)) {
-                            if (prefix != null && !setting.getName().startsWith(prefix)) {
+                            if (shouldExcludeSettingFromReset(setting, prefix)) {
                                 continue;
                             }
                             if (setting.isDefaultFromSystem()) {
@@ -3040,7 +3080,7 @@ public class SettingsProvider extends ContentProvider {
                     for (String name : settingsState.getSettingNamesLocked()) {
                         Setting setting = settingsState.getSettingLocked(name);
                         boolean someSettingChanged = false;
-                        if (prefix != null && !setting.getName().startsWith(prefix)) {
+                        if (shouldExcludeSettingFromReset(setting, prefix)) {
                             continue;
                         }
                         if (setting.isDefaultFromSystem()) {
